created player object in 'outside' room - added new room method to Player class
