+++ src/story.py
@@ -1,0 +1,29 @@
+import time
+
+def displayIntro():
+    print("welcome to the game")
+    time.sleep(1)
+    print("you win")
+    time.sleep(1)
+    print("or die")
+    time.sleep(5)
+    print("jk, you only die")
+    time.sleep(5)
+    print("yours will be nasty and awful.")
+    time.sleep(1)
+    print("will you be eaten alive by a bear?")
+    time.sleep(1)
+    print("or jump from the 50th floor to escape burning alive from the explosive jet fuel of the Boeing 767 that just slammed into the side of your office?")
+    time.sleep(5)
+    print("although, if you're lucky, a quick heart attack will knock you out while you're mowing the yard...")
+    time.sleep(4)
+    print("less lucky, you get dementia and your body slowly stops working while your family has to make the call whether to cut you loose or keep you strapped to tubes and tied down until finally one of your organs fatally fail.")
+    time.sleep(3)
+    print("you know what? nevermind. welcome to the game where you die no matter what.")
+
+# def start():
+#     player_name = input("Your name: ")
+#     time.sleep(2)
+#     answer = input("Do you want to explore? (yes/no) ")
+
+displayIntro()
